Back/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.1
cryptography==41.0.3
python-dotenv==1.0.0
gunicorn==21.2.0